requirements.txt
datasets
librosa==0.10.2.post1
huggingface_hub
hydra-core
matplotlib
numpy
numba>=0.60.0
pandas
pyfar==0.7.0
pyroomacoustics==0.8.2
PyYAML
Requests
rVADfast==0.0.3
scikit_learn
scipy
sounddevice==0.5.1
speechbrain
torch
torch_audiomentations>=0.11.1
torchaudio
torchdata
tqdm
transformers
wandb
webdataset
xlstm
